all int's are now long's. probably not the best

+++ AdventOfCode2019/Day1.cs
@@ -7,34 +7,34 @@ namespace AdventOfCode2019
 {
 	public class Day1
 	{
-		public static int Result1(string file)
+		public static long Result1(string file)
 		{
 			string[] input = File.ReadAllLines(file);
 
-			int output = 0;
+			long output = 0;
 
 			foreach (var item in input)
 			{
-				int i = (int)Math.Floor((decimal)(int.Parse(item) / 3) - 2);
+				long i = (long)Math.Floor((decimal)(long.Parse(item) / 3) - 2);
 				output += i;
 			}
 			return output;
 		}
 
-		public static int Result2(string file)
+		public static long Result2(string file)
 		{
 			string[] input = File.ReadAllLines(file);
 
-			int output = 0;
+			long output = 0;
 
 			foreach (var item in input)
 			{
-				int temp = 0;
-				int i = (int)Math.Floor((decimal)(int.Parse(item) / 3) - 2);
+				long temp = 0;
+				long i = (long)Math.Floor((decimal)(long.Parse(item) / 3) - 2);
 				temp = i;
 				while (i > 0)
 				{
-					i = (int)Math.Floor(((decimal)i / 3) - 2);
+					i = (long)Math.Floor(((decimal)i / 3) - 2);
 					if (i > 0)
 						temp += i;
 				}
