+++ AdventOfCode2019/Day8.cs
@@ -5,26 +5,26 @@ namespace AdventOfCode2019
 {
 	internal class Day8
 	{
-		internal static int Result1(string input)
+		internal static long Result1(string input)
 		{
 			string[] inputData = File.ReadAllLines(input);
 
-			int[] data = new int[inputData[0].Length];
-			for (int i = 0; i < inputData[0].Length; i++)
-				data[i] = int.Parse(new char[] { inputData[0][i] });
+			long[] data = new long[inputData[0].Length];
+			for (long i = 0; i < inputData[0].Length; i++)
+				data[i] = long.Parse(new char[] { inputData[0][(int)i] });
 
-			int height = 25;
-			int length = 6;
-			int layers = data.Length / (height * length);
+			long height = 25;
+			long length = 6;
+			long layers = data.Length / (height * length);
 
-			int[,,] image = new int[height, length, layers];
+			long[,,] image = new long[height, length, layers];
 
-			int pos = 0;
-			for (int lay = 0; lay < layers; lay++)
+			long pos = 0;
+			for (long lay = 0; lay < layers; lay++)
 			{
-				for (int y = 0; y < length; y++)
+				for (long y = 0; y < length; y++)
 				{
-					for (int x = 0; x < height; x++)
+					for (long x = 0; x < height; x++)
 					{
 						image[x, y, lay] = data[pos];
 						pos++;
@@ -32,20 +32,20 @@ namespace AdventOfCode2019
 				}
 			}
 
-			int layerLowest = 0;
-			int digit0Lowest = int.MaxValue - 1;
-			int digit1Lowest = int.MaxValue - 1;
-			int digit2Lowest = int.MaxValue - 1;
+			long layerLowest = 0;
+			long digit0Lowest = long.MaxValue - 1;
+			long digit1Lowest = long.MaxValue - 1;
+			long digit2Lowest = long.MaxValue - 1;
 
-			for (int lay = 0; lay < layers; lay++)
+			for (long lay = 0; lay < layers; lay++)
 			{
-				int digit0 = 0;
-				int digit1 = 0;
-				int digit2 = 0;
+				long digit0 = 0;
+				long digit1 = 0;
+				long digit2 = 0;
 
-				for (int y = 0; y < length; y++)
+				for (long y = 0; y < length; y++)
 				{
-					for (int x = 0; x < height; x++)
+					for (long x = 0; x < height; x++)
 					{
 						if (image[x, y, lay] == 0)
 							digit0++;
@@ -68,26 +68,26 @@ namespace AdventOfCode2019
 			return digit1Lowest * digit2Lowest;
 		}
 
-		internal static int Result2(string input)
+		internal static long Result2(string input)
 		{
 			string[] inputData = File.ReadAllLines(input);
 
-			int[] data = new int[inputData[0].Length];
-			for (int i = 0; i < inputData[0].Length; i++)
-				data[i] = int.Parse(new char[] { inputData[0][i] });
+			long[] data = new long[inputData[0].Length];
+			for (long i = 0; i < inputData[0].Length; i++)
+				data[i] = long.Parse(new char[] { inputData[0][(int)i] });
 
-			int height = 25;
-			int length = 6;
-			int layers = data.Length / (height * length);
+			long height = 25;
+			long length = 6;
+			long layers = data.Length / (height * length);
 
-			int[,,] image = new int[height, length, layers];
+			long[,,] image = new long[height, length, layers];
 
-			int pos = 0;
-			for (int lay = 0; lay < layers; lay++)
+			long pos = 0;
+			for (long lay = 0; lay < layers; lay++)
 			{
-				for (int y = 0; y < length; y++)
+				for (long y = 0; y < length; y++)
 				{
-					for (int x = 0; x < height; x++)
+					for (long x = 0; x < height; x++)
 					{
 						image[x, y, lay] = data[pos];
 						pos++;
@@ -95,17 +95,17 @@ namespace AdventOfCode2019
 				}
 			}
 
-			int[,] imageRender = new int[height, length];
-			for (int y = 0; y < length; y++)
-				for (int x = 0; x < height; x++)
+			long[,] imageRender = new long[height, length];
+			for (long y = 0; y < length; y++)
+				for (long x = 0; x < height; x++)
 					imageRender[x, y] = 9;
 
 
-			for (int y = 0; y < length; y++)
+			for (long y = 0; y < length; y++)
 			{
-				for (int x = 0; x < height; x++)
+				for (long x = 0; x < height; x++)
 				{
-					for (int lay = 0; lay < layers; lay++)
+					for (long lay = 0; lay < layers; lay++)
 					{
 						if (image[x, y, lay] < 2)
 						{
@@ -116,9 +116,9 @@ namespace AdventOfCode2019
 				}
 			}
 
-			for (int y = 0; y < length; y++)
+			for (long y = 0; y < length; y++)
 			{
-				for (int x = 0; x < height; x++)
+				for (long x = 0; x < height; x++)
 				{
 					Console.Write(imageRender[x, y]);
 				}
